- fix filetime conversion bug (yikes - this was fixed some time back in the xir300arch plugin!)

+++ plugins/xir3000/xir3000.h
@@ -200,8 +200,8 @@ typedef struct xir3000_ {
 
 // interconvert these times (ignoring leap seconds)
 
-#define FILETIME_TO_UTC(FT)  ((((FT).dwLowDateTime + (((long long) (FT).dwHighDateTime) << 32)) - UTC_FILETIME_ORIGIN_DIFF) / 1E7)
-#define UTC_TO_FILETIME(X, FT) ({* (long long *) &FT = (long long) (((X) * 1e7) + UTC_FILETIME_ORIGIN_DIFF);})
+#define FILETIME_TO_UTC(FT)  ((((unsigned)((FT).dwLowDateTime) + (((unsigned long long) (FT).dwHighDateTime) << 32)) - UTC_FILETIME_ORIGIN_DIFF) / 1E7)
+#define UTC_TO_FILETIME(X, FT) ({* (unsigned long long *) &FT = (unsigned long long) (((X) * 1e7) + UTC_FILETIME_ORIGIN_DIFF);})
 
 // a sweep is obtained in batches of pulses, and between these batches
 // the getter thread sleeps.  The following batch sizes and sleep intervals
